requirements.txt
APScheduler==3.4.0
asn1crypto==0.23.0
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
cryptography==2.1.4
diff-match-patch==20121119
Django==1.11.2
django-apscheduler==0.2.1
django-autocomplete-light==3.2.10
django-bootstrap3==9.1.0
django-braces==1.12.0
django-crispy-forms==1.7.0
django-datatable==0.3.1
django-filter==1.1.0
django-filters==0.2.1
django-import-export==0.5.1
django-tables2==1.16.0
et-xmlfile==1.0.1
idna==2.5
inflection==0.3.1
jdcal==1.3
llvmlite==0.20.0
more-itertools==4.0.1
Naked==0.1.31
ndg-httpsclient==0.4.3
numba==0.35.0
numpy==1.13.1
odfpy==1.3.5
openpyxl==2.4.9
pandas==0.20.3
pyasn1==0.4.2
pycparser==2.18
pyOpenSSL==17.5.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
Quandl==3.2.1
requests==2.17.3
six==1.10.0
tablib==0.12.1
testfixtures==5.3.0
tzlocal==1.4
unicodecsv==0.14.1
urllib3==1.21.1
xlrd==1.1.0
xlwt==1.3.0
yahoo-historical==0.1.4
